Add files via upload
Participated in a Business Case Study Competition called Brand Revive where we had been given
a real life business problem and we were supposed to find implementable solutions after
analyzing and understanding the problem well. This was the event in which we did not win but had
a turning impact on thoughts regarding my career. After this event I started to develop an
interest in Financial Analysis of Businesses and found a ground or foundation for applying my
problem solving tactics to test and solve real world business problems.